wip: framework
